chore(build): avoid circular dependencies in d.ts files
