+++ packages/core/src/index.ts
@@ -34,7 +34,7 @@ export * from './util/currentTime'
 export * from './query'
 export * from './transport'
 export * from './observable'
-export {default as Point} from './Point'
+export * from './Point'
 export {default as InfluxDB} from './InfluxDB'
 export {default as QueryApi, Row, QueryOptions} from './QueryApi'
 export {default as WriteApi} from './WriteApi'
